Source files for ResilienceApp
Resilience App written in python. As part of phase 1 ResilienceCombination.py is the server functionality that listens to ZMQ on port specified in the Dockerfile as well as in resappstart.sh. ResilenceTest.py is the test program that connects to server app through ZMQ and send and receive messages. All clients can look into ResilenceTest app as an example to see the usage pattern to  resnd and receive massages.